+++ scripts/ndfd/init_ndfd_dailyc.py
@@ -33,7 +33,7 @@ def init_year(ts):
     nc.createDimension("time", int(days))
 
     # Setup Coordinate Variables
-    lat = nc.createVariable("lat", np.float, ("lat", "lon"))
+    lat = nc.createVariable("lat", float, ("lat", "lon"))
     lat.units = "degrees_north"
     lat.long_name = "Latitude"
     lat.standard_name = "latitude"
@@ -41,7 +41,7 @@ def init_year(ts):
     lat.axis = "Y"
     lat[:] = lats
 
-    lon = nc.createVariable("lon", np.float, ("lat", "lon"))
+    lon = nc.createVariable("lon", float, ("lat", "lon"))
     lon.units = "degrees_east"
     lon.long_name = "Longitude"
     lon.standard_name = "longitude"
@@ -49,7 +49,7 @@ def init_year(ts):
     lon.axis = "X"
     lon[:] = lons
 
-    tm = nc.createVariable("time", np.float, ("time",))
+    tm = nc.createVariable("time", float, ("time",))
     tm.units = "Days since %s-01-01 00:00:0.0" % (ts.year,)
     tm.long_name = "Time"
     tm.standard_name = "time"
